Move namespace attr validator to own package
