+++ common/namespace/nsattr/errors.go
@@ -22,7 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package namespace
+package nsattr
 
 import (
 	"go.temporal.io/api/serviceerror"
